+++ includes/Webserv.hpp
@@ -1,8 +1,7 @@
 #ifndef INCLUDES_WEBSERV_HPP
 #define INCLUDES_WEBSERV_HPP
 
-#include <map>
-#include "ServerConfig.hpp"
+#include "config/ServerConfig.hpp"
 
 void server_io_multiplexing(ServerConfig &server_config);
