+++ includes/config/config.hpp
@@ -1,0 +1,8 @@
+#ifndef INCLUDES_CONFIG_CONFIG_HPP
+#define INCLUDES_CONFIG_CONFIG_HPP
+
+#include "config/ServerConfig.hpp"
+
+ServerConfig generate_server_config(const char *config_file_path);
+
+#endif /* INCLUDES_CONFIG_CONFIG_HPP */
