+++ srcs/config/ServerConfig.cpp
@@ -1,7 +1,7 @@
-#include "ServerConfig.hpp"
-#include "util.hpp"
-
 #include <algorithm>
+
+#include "config/ServerConfig.hpp"
+#include "util.hpp"
 
 #define SPACES "\v\r\f\t\n "
